requirements.txt
flanker ~= 0.4
msgpack-python ~= 0.4